requirements.txt
fastapi
uvicorn[standard]
bcrypt
pyjwt
python-dotenv
openai
EOF